Move `gardenlet` Helm chart tests to `./charts/gardener/gardenlet`
Having the test code next to the implmentation simplifies maintenance.
