+++ charts/gardener/gardenlet/.helmignore
@@ -1,6 +1,7 @@
 # Patterns to ignore when building packages.
 # This supports shell glob matching, relative path matching, and
 # negation (prefixed with !). Only one pattern per line.
+charttest
 .DS_Store
 # Common VCS dirs
 .git/
