+++ charts/gardener/gardenlet/test/test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package charttest
+package test
 
 import (
 	"context"
